refactored previous polls and new poll into polls
