--- src/routes/new-poll/new-poll.router.js
@@ -1,11 +1,0 @@
-const express = require("express");
-
-const { newPolls } = require("./new-poll.controller");
-
-const router = express.Router();
-
-router.post("", newPolls);
-
-module.exports = {
-  newPollRouter: router
-};
