--- src/routes/previous-polls/previous-polls.router.js
@@ -1,11 +1,0 @@
-const express = require("express");
-
-const { getPreviousPolls } = require("./previous-polls.controller");
-
-const router = express.Router();
-
-router.get("", getPreviousPolls);
-
-module.exports = {
-  previousPollsRouter: router
-};
